+++ GeoImage/src/main/java/org/geoimage/impl/radarsat/Radarsat2Image.java
@@ -185,7 +185,7 @@ public class Radarsat2Image extends SarImageReader {
 
 
     @Override
-    public int readPixel(int x, int y,int band) {
+    public long readPixel(int x, int y,int band) {
         TIFFImageReadParam t=new TIFFImageReadParam();
         t.setSourceRegion(new Rectangle(x, y, 1, 1));
         TIFF tiff=getImage(band);
