+++ GeoImage/src/main/java/org/geoimage/impl/tsar/TerrasarXImage.java
@@ -502,7 +502,7 @@ public class TerrasarXImage extends SarImageReader {
     }
 
     @Override
-    public int readPixel(int x, int y,int band) {
+    public long readPixel(int x, int y,int band) {
         TIFFImageReadParam t = new TIFFImageReadParam();
         t.setSourceRegion(new Rectangle(x, y, 1, 1));
         TIFF tiff=getImage(band);
